OhMyQSIM.jl Documentation

OhMyQSIM.QuantumRegister — Type

Abstract type and parent of all quantum register types

OhMyQSIM.FullStateQuantumRegister — Type

Implementation type for full state quantum register

OhMyQSIM.binary_repr — Function

binary_repr(num, N)

Get the binary string representation of a decimal integer.

OhMyQSIM.to_str — Function

to_str(qreg::FullStateQuantumRegister)

Function to convert express fullstate quantum register as a string

Examples

julia> ψ = FullStateQuantumRegister{Int}(3, "000")
julia> ψ.state[end] = 1
julia> to_str(ψ)
"(1|000>) + (1|111>)"

OhMyQSIM.apply_1qubit_full — Function

apply_1qubit_full(qreg, gate, i)

Apply the 1 qubit gate to qubit i by expanding operator fully

OhMyQSIM.apply_1qubit — Function

apply_1qubit(qreg, gate, i)

Apply the 1 qubit gate to qubit i

OhMyQSIM.apply_1qubit! — Function

apply_1qubit!(qreg, gate, i)

Apply the 1 qubit gate to qubit i inplace

OhMyQSIM.decompose_2_qubit_gate — Function

decompose_2_qubit_gate(gate; threshold=1e-15)

Decompose a 2 qubit gate into a sum of 1 qubit gates on each qubit

OhMyQSIM.apply_2qubit_full — Function

apply_2qubit_full(qreg, gate, i, j)

Apply the provided 2 qubit gate to qubits i and j by expanding full operator matrix

OhMyQSIM.swap_2qubits — Function

swap_2qubits(X)

Given a two qubit operation, switch the order of the qubits

OhMyQSIM.apply_2qubit — Function

apply_2qubit(qreg, gate, i, j)

Apply the provided 2 qubit gate to qubits i and j

OhMyQSIM.apply_2qubit! — Function

apply_2qubit!(qreg, gate, i, j)

Apply the provided 2 qubit gate to qubits i and j inplace

OhMyQSIM.measure_probs — Function

measure_probs(qreg)

Get for each bitstring

OhMyQSIM.get_conf — Function

get_conf(cprobs, num, N)

Given an array of cumulative probabilities of configurations and a random number from [0, 1], return the bitstring it corresponds to

OhMyQSIM.measure — Function

measure(qreg, shots)

Given a quantum register, return a typical set of measurements expected

OhMyQSIM.get_counts — Function

get_counts(results)

Given a set of measurement results in the form of bitstrings, return the counts of each